Reconcile differences from Soundtracker_v1-v9 docs
I was cross-checking this information with the "Soundtracker_v1-v9"
archive (which I believe I got from Aminet) containing binaries and
documentation of different Soundtracker versions but not taking the 3
and 4 fork, so it follows Il Scuro's 3 and AFL 4 instead of DOC's.

In AFL 4 it claims:
History     : SoundTracker V4.0 is based on the original V1.21
              version by Karsten Obarski. Later edited by the
              Exterminator, then completed by IL SCURO and last by DOC
              with their Equalizer, but without TEMPO setting!!!

Which suggests that DOC 3 has an equalizer, but is not clear about AFL
4 having or not this feature. This could be a good marker to decide if
AFL descends from DOC 3 or Il Scuro's 3.

Also looking at the language in these files, it seems that "X is based
on Y" is frequently used in the context of lineage and not direct
descent, and in this case DOC 4 and 6 would have a more normal
based-on-previous-version relationship instead of being direct
descendants of TJC 2.

Note however that these documents are not completely reliable since
they contain documentation (and binaries!) for what would be DOC 5
(which makes me believe that the author of the documentation didn't
know exactly what happened, but he has complete d-m-y dates for most
releases). I didn't check this binary yet to see what it contains. I'm
not sure if you have this archive or not, so I'll send it attached
just in case.

I tweaked the graph sources to reflect this new view and added
comments to the txt files. We can tweak it further and conduct some
additional research until we reach consensus on the most historically
accurate layout.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>